Web App and MongoDB plus Web App and SQL Database

+++ samples/web-app-cosmosdb-mongodb-api/python/bicep/deploy.sh
@@ -153,4 +153,6 @@ else
 fi
 
 # Remove the zip package of the web app
-rm "$ZIPFILE"
+if [ -f "$ZIPFILE" ]; then
+	rm "$ZIPFILE"
+fi
